+++ src/bfd/elf32-openrisc.c
@@ -347,8 +347,8 @@ openrisc_elf_relocate_section (bfd *output_bfd,
 
 	  name =
             bfd_elf_string_from_elf_section(input_bfd,
-                                            symtab_hdr->sh_link,
-                                            sym->st_name);
+                                            (unsigned int)symtab_hdr->sh_link,
+                                            (unsigned int)sym->st_name);
 	  name = ((name == NULL) ? bfd_section_name(input_bfd, sec) : name);
 	}
       else
